pyproject.toml
[tool.poetry]
name = "examplesite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
ipython = "^8.14.0"
psycopg2-binary = "^2.9.7"
django-registration = "^3.4"
django-seed = "^0.3.1"
django-ckeditor = "^6.7.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
drf-yasg = "^1.21.7"
freezegun = "^1.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
mypy = "^1.6.1"
time-machine = "^2.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
